Move static files into module directory
To be able to distribute the package, we need to include the static word lists
into the package module.

Also, add MANIFEST.in to tell setup.py that it's able to include the files. The
MANIFEST.in file is generic and any new added files into xkcdpass/static/ will
be added into new versions of the package.