"Fix" tests
Due to the changes incurred by the packaging, the tests needed some changes as
well. I've fixed those, but three tests still fail. However, I ran the tests on
origin/master, and three of them were failing.
I can help with that, but it's out of scope for this PR. ;)