Add a wordlist based on the wiktionary
The EFF wordlist has a lot of awesome properties, like no word being a
prefix of another, but is designed for use with dices, so it is a bit
short.

Building a wordlist out of the words defined in dictionary (like the
legacy one was) creates a much bigger list, but with very obscure
words.

Most lists of common words (like the Oxford English Corpus) are not
freely available, so I thought of finding the most common words in
Wikipedia.

Unfortunately for this, I found wikipedia is a bit too big and hard to
process.

The second best was the wiktionary. The idea is not to include every
word defined in it, but to find the common words used in definitions
and example.

Processing the wiktionary is made easy by the json published by
https://kaikki.org.

The wordlist in this PR was created by an odd mix of shell and go. The
main script is

```
set -euo pipefail
export LANG=c

curl -O https://kaikki.org/dictionary/English/kaikki.org-dictionary-English.json

cat kaikki.org-dictionary-English.json |
    jq ' .senses | map(.glosses,(.examples //[] | map(.text)))' |
    tr -c '[:alpha:]' '\n'          | # turn non alpha into newlines
    sed -e 's/vv/w/g' -e 's/VV/W/g' | # OCR issue I guess
    grep  '...'                     | # at least 3 characters
    grep -v '.........'             | # at most 8 characters
    grep '[a-z]'                    | # at least a lowercase letter
    tr '[:upper:]' '[:lower:]'      | # ignore case
    sort | uniq -c | sort -n > common

grep -E -v ' (1|2|3|4|5|6|7|8|9|10|11|12|13|14|15|16|17|18|19|20|21|22|23|24|25|26|27|28) ' common | # only common words
    awk '{print $2}'    | # only the word
    sort | go run remove-prefixes.go > wiktionary
```

And remove-prefixes.go removes words so that no word is a prefix of another:

```
package main

import (
	"os"
	"strings"
	"fmt"
	"bufio"
)

func readLines() []string {
	s := bufio.NewScanner(os.Stdin)
	ret := []string{}
	for s.Scan() {
		ret = append(ret, s.Text())
	}
	return ret
}

func prefixOf(lines []string, i int) int {
	l := lines[i]
	n := 0
	for _,o := range(lines[i+1:]) {
		if strings.HasPrefix(o, l) {
			n += 1
		} else {
			return n
		}
	}
	return n
}

func remove(v []string, i int) []string {
	return append(v[0:i], v[i+1:]...)
}

func main() {
	lines := readLines()
	for i := len(lines) - 1; i >= 0; i-- {
		n := prefixOf(lines, i)
		if n > 1 {
			lines = remove(lines, i)
		} else if n == 1 {
			lines = remove(lines, i + 1)
		}
	}

	for _,v := range(lines) {
		fmt.Println(v)
	}
}
```

There is a lot of room for yak shaving over the exact heuristics, but
a quick sampling with

xkcdpass -w wiktionary --min 3 --max 8

creates passwords that I find easy to remember. The list also has over
2x as many words as eff-long and the average word size is smaller.

I can convert the generation script from sh+go to python and include
it in the pull request if desired.